+++ CHANGELOG.md
@@ -25,6 +25,7 @@
   - Added support for Ubuntu 26.04+
   - Enforced minimum required version 26.04 (exits with error on < 26.04)
   - Added `distros/ubuntu/setup.sh` and `distros/ubuntu/packages.sh` with ButterRepo repository configuration
+    - Which caused conflict in quickshell and mangowc
   - Updated noctalia-greeter Polkit rule to permit `sudo` group for Ubuntu/Debian
 - Arch Install
   - Fixed packages to be installed